test(rollout): model migration after blue-green cutover
