+++ test-suite/fhevm/rollouts/v0.14-to-v0.15-gpu-key-migration/versions.ts
@@ -1,13 +1,5 @@
 type Env = Record<string, string | undefined>;
 
-const required = (env: Env, name: string): string => {
-  const value = env[name]?.trim();
-  if (!value) {
-    throw new Error(`${name} is required and must be an exact image tag or commit SHA`);
-  }
-  return value;
-};
-
 export const connectorVersionKeys = [
   "CONNECTOR_DB_MIGRATION_VERSION",
   "CONNECTOR_GW_LISTENER_VERSION",
@@ -30,7 +22,6 @@ export const coprocessorVersionKeys = [
 export type MigrationVersions = {
   baseline: Record<string, string>;
   baselineTag: string;
-  blueTag: string;
 };
 
 const requiredVersion = (versions: Record<string, string>, key: string): string => {
@@ -42,7 +33,6 @@ const requiredVersion = (versions: Record<string, string>, key: string): string 
 export const migrationPhaseVersions = (
   baseline: Record<string, string>,
   target: Record<string, string>,
-  blueTag: string,
 ) => {
   const contract: Record<string, string> = {
     ...baseline,
@@ -53,29 +43,17 @@ export const migrationPhaseVersions = (
     CORE_VERSION: requiredVersion(target, "CORE_VERSION"),
     ...Object.fromEntries(connectorVersionKeys.map((key) => [key, requiredVersion(target, key)])),
   };
-  const blue: Record<string, string> = {
-    ...connector,
-    ...Object.fromEntries(coprocessorVersionKeys.map((key) => [key, blueTag])),
-  };
-  return { blue, contract, connector };
+  return { contract, connector };
 };
 
 export const migrationVersions = (env: Env = process.env): MigrationVersions => {
   const releaseTag = env.RFC029_BASELINE_FHEVM_TAG?.trim() || "v0.14.0-10";
   // v0.14.0-10 did not publish a host-contracts image; that component remained on v0.14.0-9.
   const hostTag = env.RFC029_BASELINE_HOST_TAG?.trim() || "v0.14.0-9";
-  const requestedBlueTag = required(env, "RFC029_BLUE_TAG");
-  if (!/^(?:v0\.15\.0-\d+|[0-9a-f]{7}|[0-9a-f]{40})$/i.test(requestedBlueTag)) {
-    throw new Error("RFC029_BLUE_TAG must be a v0.15.0-N release tag or an exact commit SHA");
-  }
-  const blueTag = /^[0-9a-f]{40}$/i.test(requestedBlueTag)
-    ? requestedBlueTag.slice(0, 7)
-    : requestedBlueTag;
   const kmsCoreTag = env.RFC029_KMS_CORE_TAG?.trim() || "v0.14.0-1";
 
   return {
     baselineTag: releaseTag,
-    blueTag,
     baseline: {
       GATEWAY_VERSION: releaseTag,
       HOST_VERSION: hostTag,
@@ -87,6 +65,6 @@ export const migrationVersions = (env: Env = process.env): MigrationVersions => 
 };
 
 export const versionSources = [
-  "rollout=v0.15-to-v0.15.1-gpu-key-migration",
+  "rollout=v0.14-to-v0.15-gpu-key-migration",
   "migration=same-key-compressed-xof",
 ];
